Fixes exponential code explosion bug
There were two issues.  First, the order of operations in the dbcoil flux calculation was off from the intent of the IDL code.  Second, the array in make_reduced_array was not dimensioned correctly, so it wasn't getting the proper values.  Another small issue was found and fixed.  The flux returned needed to be divided by 2 pi and the reduced matrix calculation was missing a factor of 1e6