Bug fix of move_can
The main issue was that the dbcoilbz function was not dependent on the current in the coil.  Once this was fixed, the results look just about like the results from the AJP article.  You still can't run further than 70 microseconds because the can has split in two by this point.