Adds temperature setting
Adds a method to set the temperature of the driven coils, (the can or sample).  Removes the last of the old script driven functions.  Also contains examples of how to add a legend and axis titles.  There is a sim that compares the driving current at room temperature to the current at liquid helium temperature.